Update README.md
When building from scratch using 2017-07-05-raspbian-jessie I get errors "Unable to locate package php" and errors on all php packages. The current php7.0 is not found, but php5 is. Sorry that my earlier suggestion that "libcppunit-1.13-0" should change to "libcppunit-1.13-0v5" was incorrect for Jessie, I was attempting to use Raspbian-stretch.